requirements.txt
# ================================================================
# Zero-Shot Clustering for Animal Images - Requirements
# ================================================================
# Tested with Python 3.10+
# Install: pip install -r requirements.txt

# ----------------------------------------------------------------
# PyTorch (with CUDA 12.8 support - adjust for your system)
# ----------------------------------------------------------------
--extra-index-url https://download.pytorch.org/whl/cu128
torch>=2.0.0
torchvision>=0.15.0

# ----------------------------------------------------------------
# Core ML Libraries
# ----------------------------------------------------------------
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# ----------------------------------------------------------------
# Dataset Download
# ----------------------------------------------------------------
datasets>=2.14.0           # HuggingFace Datasets
huggingface_hub>=0.20.0    # HuggingFace Hub for authentication
pyarrow>=10.0.0            # Parquet file reading

# ----------------------------------------------------------------
# Dimension Reduction
# ----------------------------------------------------------------
umap-learn>=0.5.5          # UMAP dimension reduction

# ----------------------------------------------------------------
# Clustering
# ----------------------------------------------------------------
hdbscan>=0.8.33            # Hierarchical DBSCAN clustering

# ----------------------------------------------------------------
# Visualization
# ----------------------------------------------------------------
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0

# ----------------------------------------------------------------
# Vision Transformer Models
# ----------------------------------------------------------------
# DINOv2 & DINOv3
timm>=0.9.0                # PyTorch Image Models (for DINOv2)
transformers>=4.35.0       # HuggingFace Transformers (for DINOv3, SigLIP)
safetensors>=0.4.0

# CLIP
git+https://github.com/openai/CLIP.git
ftfy>=6.1.0
regex>=2023.0.0

# BioCLIP 2 (OpenCLIP)
open_clip_torch>=2.23.0

# ----------------------------------------------------------------
# Image Processing
# ----------------------------------------------------------------
pillow>=10.0.0

# Note: GUI uses Tkinter which is built into Python (no extra install needed)
